backend/pyproject.toml
[project]
name = "cms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dhwanil Shah", email = "itzdhwanil@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "granian>=2.3.1",
    "orjson>=3.10.18",
    "phonenumbers>=9.0.7",
    "pydantic>=2.11.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "structlog>=25.3.0",
    "supabase>=2.16.0",
    "uuid-utils>=0.11.0",
    "uvloop>=0.21.0",
]

[project.scripts]
server = "cms.app.server:start_server"
migrations = "cms.utils.migrations:run_cli"

[build-system]
requires = ["uv_build>=0.7.3,<0.8"]
build-backend = "uv_build"


[tool.uv.build-backend]
module-name = "cms"
module-root = ""

[dependency-groups]
dev = [
    "granian[reload]>=2.3.1",
]

